feature: AC-319 improve NGiNX reverse proxying (#48)
* Consolidate into a single NGiNX configuration

* Add pgadmin4 and keycloak to NGiNX

* Remove extra slash
Set both localhost and nginx as valid server names

* Improve cookie security

* Better explain why we can't use HttpOnly here

* Improve error handling when first connecting to the OIDC provider

* Stop mapping the Keycloak port

* fix a bug that was causing keycloak to be shut down unexpectedly

* Fix a variable bug in the connection strings

* Update http files to use the new consolidated NGiNX urls

* Note about trusting the certificate

* Another rename of "Starting Blocks"

* Update generate-certificate.sh to improve subject handling for certificate generation

* Correction on env var usage

+++ compose/down.ps1
@@ -22,10 +22,12 @@ else {
 }
 
 if ($Keycloak) {
+  if ($V) {
     # Shut down Keycloak and remove its volume
     docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down --volumes
-}
-else {
-    # Shut down Keycloak without removing its volume
-    docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down
+  }
+  else {
+      # Shut down Keycloak without removing its volume
+      docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down
+  }
 }
